requirements.txt
fastapi
uvicorn
scikit-learn==1.7.2
pandas
joblib
python-dotenv
requests
pydantic
